turn float32 to float16(shl , no matter)

+++ SHL-2024/gen_rfft.py
@@ -26,7 +26,7 @@ def window_rfft(tmp, window_length=100, step=20, fs=100):
 
 def gen_rfft(dataset, shape):
     source = np.load('data/npy_data/{}/data.npy'.format(dataset),mmap_mode='r')
-    data = open_memmap('data/fft_data/{}/data.npy'.format(dataset),dtype=np.float32,mode='w+',shape=shape)
+    data = open_memmap('data/fft_data/{}/data.npy'.format(dataset),dtype=np.float16,mode='w+',shape=shape)
     for i in tqdm(range(shape[0])):
         Parallel(n_jobs=psutil.cpu_count(logical=False), verbose=0)(delayed(lambda j: data[i].__setitem__(j,window_rfft(source[i,j])))(j) for j in range(shape[1]))
 
@@ -35,7 +35,7 @@ def gen_test(shape=(9, 92726, 1071)):
         os.mkdir('data/fft_data/test')
     #生成数据
     source = np.load('data/npy_data/test/data.npy',mmap_mode='r')
-    data = open_memmap('data/fft_data/test/data.npy',dtype=np.float32,mode='w+',shape=shape)
+    data = open_memmap('data/fft_data/test/data.npy',dtype=np.float16,mode='w+',shape=shape)
 
     Parallel(n_jobs=psutil.cpu_count(logical=False), verbose=0)(delayed(lambda i: data.__setitem__(i,window_rfft(source[i])))(i) for i in tqdm(range(shape[0])))
